move car-ml parts to car-app
