+++ car-app/xebikart/parts/rl.py
@@ -1,4 +1,3 @@
-import tensorflow as tf
 import numpy as np
 
 from xebikart.rl.sac import CustomSAC
