--- car-ml/xebikart/parts/__init__.py
@@ -1,5 +1,0 @@
-import tensorflow as tf
-
-
-def enable_eager_execution():
-    tf.compat.v1.enable_eager_execution()
